Found the issue: - PETITMODBUS_TIMEOUTTIMER must be set to the minimum of 3.5 char timeout time (as still on 250ms)!!! - fixed some other stuff - set the master on tmr2 PR2 activation on TMR2 roll over --> 3 slaves 10.90ms @115.200kbps - todo: readback the slave communication

+++ SiebwaldeTrackAmplifier2.X/modbus/General.c
@@ -1,3 +1,3 @@
 #include "General.h"
 
-volatile unsigned short Timer1_Tick_Counter;
+volatile unsigned int Timer1_Tick_Counter;
